Add confirmation window every time a user deletes a person from the list

+++ src/App.jsx
@@ -64,12 +64,14 @@ const App = () => {
     return isDuplicate
   }
 
-  const deletePersonFromList = (id) => {
-    phonebookService
+  const deletePersonFromList = (name, id) => {
+    if (window.confirm(`Delete ${name}?`)) {
+      phonebookService
       .deleteName(id)
       .then(deletedPerson => {
         setPersons(persons.filter(person => person.id !== deletedPerson.id))
       })
+    }
   }
 
   const namesToShow = filter
